Ibaldachini/fixes and linux

+++ Linux/AT28CProgrammer/AT28CProgrammer.c
@@ -27,8 +27,8 @@ int requestRead(int fd, e_rom_type romtype);
 // invia il comando di richiesta scrittura della memoria
 int requestWrite(int fd, e_rom_type romtype, bool paged);
 
-// legge e visualizza la risposta dal programmatore
-int readAnswer(int fd, long msec);
+// legge e visualizza o salva nel buffer la risposta dal programmatore
+int readAnswer(int fd, char* buffer, long msec);
 
 // legge la risposta dal programmatore con il contenuto della memoria e lo salva sul file indicato, attende la risposta per max msec millisecondi 
 int readEprom(int fd, e_rom_type romtype, char* filename, long msec);
@@ -42,11 +42,11 @@ int writeEprom(int fd, e_rom_type romtype, bool paged, char* filename, long msec
 // setup Software Data Protection
 int setupSDP(int fd, bool enable, long msec);
 
-// invia al programmatore il byte da scrivere, attende al massimo msecforbyte millisecondi
-int writeByte(int fd, int address, unsigned char val, long msecforbyte);
+// invia al programmatore la locazione di memoria e il byte da scrivere
+int requestWriteByte(int fd, int address, unsigned char val, long msecforbyte);
 
-// legge dal programmatore la locazione di memoria richiesta, attende al massimo msecforbyte millisecondi
-int readByte(int fd, int address, long msecforbyte);
+// legge al programmatore la locazione di memoria da leggere
+int requestReadByte(int fd, int address, long msecforbyte);
 
 // applicazione principale
 int main (int argc, char **argv) {
@@ -59,6 +59,9 @@ int main (int argc, char **argv) {
   // indicatore operazione paginata
   bool paged = false;
 
+  // indicatore leura scrittura singolo byte
+  bool singlebyte = false;
+
   // nome del device seriale a cui è collegato il programmatore
   char *device = NULL;
 
@@ -69,7 +72,7 @@ int main (int argc, char **argv) {
   int address = -1;
 
   // valore da scrivere
-  int val = 256;
+  int val = -1;
 
   // effettua il parsing dei parametri passati da linea di comando
   int c;
@@ -85,10 +88,10 @@ int main (int argc, char **argv) {
         break;
       // selezione tipologia di memoria
       case 't':
-        romtype = atoi(optarg);
-        if (romtype < 0 || romtype >= NONE) {
+        if (strcmp("AT28C64", optarg) == 0) romtype = AT28C64;
+        if (strcmp("AT28C256", optarg) == 0) romtype = AT28C256;
+        if (romtype == NONE) {
           printf("unknown romtype\n");
-          return -1;
         }
         break;
       case 'o':
@@ -99,12 +102,20 @@ int main (int argc, char **argv) {
           if (optarg[1] == 'p') {
             paged = true;
           }
+          // opzione per la scrittura di un byte
+          if (optarg[1] == 'b') {
+            singlebyte = true;
+          }
         // opzione per la verifica della memoria
         } else if (optarg[0] == 'v') {
           operation = 'v';
         // opzione per la lettura della memoria
         } else if (optarg[0] == 'r') {
           operation = 'r';
+          // opzione per la lettura di un byte
+          if (optarg[1] == 'b') {
+            singlebyte = true;
+          }
         // opzione per l'abilitazione del software data protection
         } else if (optarg[0] == 'e') {
           operation = 'e';
@@ -114,16 +125,25 @@ int main (int argc, char **argv) {
         }
         else {
           printf("unknown operation\n");
-          return -1;
         }
         break;
       // indirizzo da leggere o scrivere
       case 'a':
-        address = atoi(optarg);
+        if (optarg[0] == 'x') {
+          sscanf(optarg + 1, "%x", &address);
+        } else {
+          address = atoi(optarg);
+        }
         break;
       // valore da scrivere
       case 'b':
-        val = atoi(optarg);
+        if (optarg[0] == 'x') {
+          int ival;
+          sscanf(optarg + 1, "%x", &ival);
+          val = ival;
+        } else {
+          val = atoi(optarg);
+        }
         if (val > 255) {
           printf("wrong value\n");
           return -1;
@@ -134,30 +154,41 @@ int main (int argc, char **argv) {
 
   if (address >= 8192 && romtype == AT28C64) {
     printf("wrong address\n");
-    return -1;
+    address = -1; // print help if needed
   }
 
   if (address >= 32768 && romtype == AT28C256) {
     printf("wrong address\n");
-    return -1;
+    address = -1; // print help if needed
   }
 
   // se non sono stati impostati gli argomenti obbligatori visualizza l'help ed esce
   if (device == NULL ||
-      (filename == NULL && (operation == 'w' || operation == 'r' || operation == 'v')) ||
       operation == 0 ||
-      romtype == NONE) {
-    printf("use: AT28CProgrammer -d <device> [-f <filename>] -o <operation> -t <romtype>\n");
+      romtype == NONE ||
+      (filename == NULL && (operation == 'w' || operation == 'v') && singlebyte == false) ||
+      (address == -1 && (operation == 'w' || operation == 'r') && singlebyte == true) ||
+      (val == -1 && operation == 'w' && singlebyte == true)) {
+    printf("AT28CProgrammer V.1.01\n");
+    printf("use: AT28CProgrammer -d <device> -t <romtype> -o <operation> [-a <address>] [-b <byte>] [-f <filename>]\n");
     printf("\t-d: serial port\n");
-    printf("\t-f: file name to read or write\n");
-    printf("\t-o r: set to read eprom\n");
+    printf("\t-t AT28C64: eeprom type AT28C64\n");
+    printf("\t-t AT28C256: eeprom type AT28C256\n");
+    printf("\t-o r: set to read eprom (save to file or dump to screen if no file selected)\n");
+    printf("\t-o rb: set to read byte (needed -a parameter)\n");
     printf("\t-o w: set to write eprom\n");
+    printf("\t-o wp: set to paged write eprom (only supported by AT28C256)\n");
+    printf("\t-o wb: set to write byte (needed -a and -b parameters)\n");
     printf("\t-o v: set to verify eprom\n");
     printf("\t-o e: set to enable software data protection\n");
     printf("\t-o d: set to disable software data protection\n");
-    printf("\t-t: rom type 0 = AT28C64, 1 = AT28C256\n");
-    printf("\texample: AT28CProgrammer -d /dev/ttyUSB0 -f /tmp/dump.bin -t 1 -o r\n");
-    printf("\texample: AT28CProgrammer -d /dev/ttyUSB0 -f /tmp/towrite.bin -t 0 -o w\n");
+    printf("\t-a: address to read or write for single byte mode (decimal or preceded with x for hex)\n");
+    printf("\t-b: byte to write for single byte mode (decimal or preceded with x for hex)\n");
+    printf("\t-f: file name to read or write\n");
+    printf("read  example:      AT28CProgrammer -d /dev/ttyUSB0 -t AT28C256 -o r -f /tmp/dump.bin\n");
+    printf("write example:      AT28CProgrammer -d /dev/ttyUSB0 -t AT28C64 -o w -f /tmp/towrite.bin\n");
+    printf("read byte example:  AT28CProgrammer -d /dev/ttyUSB0 -t AT28C64 -o rb -a 4096\n");
+    printf("read byte example:  AT28CProgrammer -d /dev/ttyUSB0 -t AT28C64 -o rb -a x1000\n");
     return -1;
   }
 
@@ -176,12 +207,12 @@ int main (int argc, char **argv) {
   }
 
   // impostazione dei parametri della porta seriale
-  struct termios tio; 
-  if (tcgetattr(fd, &tio) != 0) { 
+  struct termios tio;
+  if (tcgetattr(fd, &tio) != 0) {
     close(fd);
-    printf("error tcgetattr\n"); 
-    return -1; 
-  } 
+    printf("error tcgetattr\n");
+    return -1;
+  }
 
   // minimum number of character for non canonical read
   tio.c_cc[VMIN] = 1;
@@ -203,12 +234,12 @@ int main (int argc, char **argv) {
   cfsetispeed(&tio, B115200);
   cfsetospeed(&tio, B115200);
 
-  // apply settings to serial port 
-  if (tcsetattr(fd, TCSANOW, &tio) != 0) { 
+  // apply settings to serial port
+  if (tcsetattr(fd, TCSANOW, &tio) != 0) {
     close(fd);
-    printf("error tcsetattr\n"); 
-    return -1; 
-  } 
+    printf("error tcsetattr\n");
+    return -1;
+  }
 
   // imposta la seriale in modalità non blocking durante la lettura e pulisce eventuali comunicazioni residue
   fcntl(fd, F_SETFL, FNDELAY);
@@ -223,10 +254,10 @@ int main (int argc, char **argv) {
   }
 
   // attende la risposta per un massimo di 100 ms (se il dispositivo non è stato resettato nell'apertura della comunicazione risponderà qui)
-  if (readAnswer(fd, 100) == -1) {
-    // non ha ricevuto la risposta alla versione firmware  
+  if (readAnswer(fd, NULL, 100) == -1) {
+    // non ha ricevuto la risposta alla versione firmware
     // attende l'eventuale intestazione inviata dal programmatore per massimo 1.5 secondi
-    if (readAnswer(fd, 1500) == -1) {
+    if (readAnswer(fd, NULL, 1500) == -1) {
       close(fd);
       printf("error reading header\n");
       return -1;
@@ -240,7 +271,7 @@ int main (int argc, char **argv) {
     }
 
     // attende la risposta contentente la versione firmware per un massimo di 100 ms
-    if (readAnswer(fd, 100) == -1) {
+    if (readAnswer(fd, NULL, 100) == -1) {
       close(fd);
       printf("error reading firmware version\n");
       return -1;
@@ -264,12 +295,29 @@ int main (int argc, char **argv) {
   }
   // verifica se richiesta scrittura della memoria
   else if (operation == 'w') {
-    if (address != -1) {
+    if (singlebyte) {
       // invia il comando di richiesta scrittura della byte
-      if (writeByte(fd, address, val, 100) == -1) {
+      if (requestWriteByte(fd, address, val, 100) == -1) {
         close(fd);
         printf("error request write byte\n");
         return -1;
+      }
+      // attende la risposta contentente il byte presente nella EPROM dopo la scrittura
+      char buffer[64];
+      if (readAnswer(fd, buffer, 100) == -1) {
+        close(fd);
+        printf("error reading firmware version\n");
+        return -1;
+      }
+      if (memcmp(buffer, "+WRITEBYTE=", 11) == 0) {
+        int wval;
+        sscanf(buffer + 11, "%d", &wval);
+        unsigned char c = (unsigned char)wval;
+        if (c != val) {
+          printf("write error, read byte %u [x%02X] at address %u [x%04X]\n", c, c, (unsigned int)address, (unsigned int)address);
+        } else {
+          printf("written byte %u [x%02X] at address %u [x%04X]\n", c, c, (unsigned int)address, (unsigned int)address);
+        }
       }
     } else {
       // invia il comando di richiesta scrittura della memoria selezionata
@@ -288,12 +336,25 @@ int main (int argc, char **argv) {
   }
   // verifica se richiesta lettura della memoria
   else if (operation == 'r') {
-    if (address != -1) {
+    if (singlebyte) {
       // invia il comando di richiesta scrittura della byte
-      if (readByte(fd, address, 100) == -1) {
+      if (requestReadByte(fd, address, 100) == -1) {
         close(fd);
         printf("error request read byte\n");
         return -1;
+      }
+      // attende la risposta contentente il byte letto dalla EPROM
+      char buffer[64];
+      if (readAnswer(fd, buffer, 100) == -1) {
+        close(fd);
+        printf("error reading firmware version\n");
+        return -1;
+      }
+      if (memcmp(buffer, "+READBYTE=", 10) == 0) {
+        int val;
+        sscanf(buffer + 10, "%d", &val);
+        unsigned char c = (unsigned char)val;
+        printf("read byte %u [x%02X] at address %u [x%04X]\n", c, c, (unsigned int)address, (unsigned int)address);
       }
     } else {
       // invia il comando di richiesta lettura della memoria selezionata
@@ -380,8 +441,8 @@ int requestWrite(int fd, e_rom_type romtype, bool paged) {
   return -1;
 }
 
-// legge e visualizza la risposta dal programmatore
-int readAnswer(int fd, long msec) {
+// legge e visualizza o salva nel buffer la risposta dal programmatore
+int readAnswer(int fd, char* buffer, long msec) {
   int readed = 0;
   while (true) {
     fd_set rfds;
@@ -402,11 +463,16 @@ int readAnswer(int fd, long msec) {
       // ricevuto risposta legge 1 carattere e lo stampa
       char c;
       read(fd, &c, 1);
-      readed++;
       // filtra i caratteri di line feed e carriage return
       if (c != '\n' && c != '\r') {
-        printf("%c", c);
+        if (buffer == NULL) {
+          printf("%c", c);
+        } else {
+          buffer[readed] = c;
+          buffer[readed + 1] = 0;
+        }
       }
+      readed++;
       // ha ricevuto almeno un carattere,
       // imposta il timeout per i prossimi caratteri a 100 ms
       msec = 100;
@@ -418,7 +484,9 @@ int readAnswer(int fd, long msec) {
 
   // se ha ricevuto dei caratteri solo alla fine visualizza un fine riga
   if (readed) {
-    printf("\n");
+    if (buffer == NULL) {
+      printf("\n");
+    }
     return 0;
   }
 
@@ -438,12 +506,16 @@ int readEprom(int fd, e_rom_type romtype, char* filename, long msec) {
   if (romtype == AT28C256) {
     totalbytes = 32768;
   }
-  unlink(filename);
-  writefd = open(filename, O_WRONLY | O_CREAT | O_TRUNC, S_IRUSR | S_IWUSR | S_IRGRP | S_IROTH);
-  if (writefd == -1) {
-    printf("error opening output file\n");
-    return -1;
+  if (filename != NULL) {
+    unlink(filename);
+    writefd = open(filename, O_WRONLY | O_CREAT | O_TRUNC, S_IRUSR | S_IWUSR | S_IRGRP | S_IROTH);
+    if (writefd == -1) {
+      printf("error opening output file\n");
+      return -1;
+    }
   }
+  char buf[128];
+  char bufr[16 + 1];
   while (true) {
     fd_set rfds;
     struct timeval tv;
@@ -460,13 +532,28 @@ int readEprom(int fd, e_rom_type romtype, char* filename, long msec) {
       printf("error select\n");
       return -1;
     } else if (retval > 0) {
-      char c;
+      unsigned char c;
       read(fd, &c, 1);
+      if (filename != NULL) {
+        write(writefd, &c, 1);
+      } else {
+        if (readed % 16 == 0) {
+          sprintf(buf, "x%04X: ", readed);
+        }
+        sprintf(buf + strlen(buf), " x%02X", c);
+        bufr[readed % 16] = c >= 0x20 && c < 0x7F ? c : '.';
+      }
       readed++;
-      write(writefd, &c, 1);
+      if (filename == NULL && readed % 16 == 0) {
+        bufr[16] = 0;
+        sprintf(buf + strlen(buf), "  -  | %s |\n", bufr);
+        printf(buf);
+      }
       int perc = readed * 100 / totalbytes;
       if (perc != lastperc) {
-        printf("<- read percent: %d%%\r", perc);
+        if (filename != NULL) {
+          printf("<- read percent: %d%%\r", perc);
+        }
         fflush(stdout);
         lastperc = perc;
       }
@@ -476,13 +563,15 @@ int readEprom(int fd, e_rom_type romtype, char* filename, long msec) {
     }
   }
 
-  if (readed) {
+  if (filename != NULL && readed) {
     printf("\n");
   }
 
   // visualizza il numero di bytes ricevuti
   printf("read: %d\n", readed);
-  close(writefd);
+  if (filename != NULL) {
+    close(writefd);
+  }
 
   // verifica se ha ricevuto il numero di bytes attesi
   if (readed != totalbytes) {
@@ -563,7 +652,7 @@ int verifyEprom(int fd, e_rom_type romtype, char* filename, long msec) {
   if (readed != totalbytes) {
     return -1;
   }
-  
+
   if (errors) {
     printf("%d errors found\n", errors);
     return -1;
@@ -624,7 +713,7 @@ int writeEprom(int fd, e_rom_type romtype, bool paged, char* filename, long msec
     bool err = false;
     for (size_t idx = 0; idx < blocksize; idx++) {
       if (rbuf[idx] != buf[idx]) {
-        printf("\n-> written byte: 0x%02X, read byte: 0x%02X\n", buf[idx], rbuf[idx]);
+        printf("\n-> written byte: %u [x%02X], read byte: %u [x%02X]\n", buf[idx], buf[idx], rbuf[idx], rbuf[idx]);
         err = true;
         break;
       }
@@ -675,73 +764,24 @@ int setupSDP(int fd, bool enable, long msec) {
   return -1;
 }
 
-// invia al programmatore il byte da scrivere, attende al massimo msecforbyte millisecondi
-int writeByte(int fd, int address, unsigned char val, long msecforbyte) {
+// invia al programmatore la locazione di memoria e il byte da scrivere
+int requestWriteByte(int fd, int address, unsigned char val, long msecforbyte) {
   tcflush(fd, TCIOFLUSH);
 
   const char* cmdWriteByte = "WRITEBYTE=%d,%d\r";
   char buff[32];
   sprintf(buff, cmdWriteByte, address, val);
-  printf("write byte 0x%02X at address 0x%04X\n", (unsigned char)val, (unsigned int)address);
-  write(fd, cmdWriteByte, strlen(cmdWriteByte));
-
-  fd_set rfds;
-  struct timeval tv;
-  int retval;
-
-  FD_ZERO(&rfds);
-  FD_SET(fd, &rfds);
-
-  tv.tv_sec = (msecforbyte * 1000) / 1000000;
-  tv.tv_usec = (msecforbyte * 1000) % 1000000;
-
-  retval = select(fd + 1, &rfds, NULL, NULL, &tv);
-  if (retval == -1) {
-    printf("error select\n");
-    return -1;
-  } else if (retval > 0) {
-    // ricevuto risposta legge 1 carattere e lo stampa
-    char c;
-    read(fd, &c, 1);
-    printf("written byte 0x%02X at address 0x%04X\n", (unsigned char)c, (unsigned int)address);
-  } else {
-    return -1;
-  }
-
-  return 0;
+  printf("write byte %u [x%02X] at address %u [x%04X]\n", (unsigned char)val, (unsigned char)val, (unsigned int)address, (unsigned int)address);
+  return write(fd, buff, strlen(buff));
 }
 
-// legge dal programmatore la locazione di memoria richiesta, attende al massimo msecforbyte millisecondi
-int readByte(int fd, int address, long msecforbyte) {
+// legge al programmatore la locazione di memoria da leggere
+int requestReadByte(int fd, int address, long msecforbyte) {
   tcflush(fd, TCIOFLUSH);
 
   const char* cmdReadByte = "READBYTE=%d\r";
   char buff[32];
   sprintf(buff, cmdReadByte, address);
-  write(fd, cmdReadByte, strlen(cmdReadByte));
-
-  fd_set rfds;
-  struct timeval tv;
-  int retval;
-
-  FD_ZERO(&rfds);
-  FD_SET(fd, &rfds);
-
-  tv.tv_sec = (msecforbyte * 1000) / 1000000;
-  tv.tv_usec = (msecforbyte * 1000) % 1000000;
-
-  retval = select(fd + 1, &rfds, NULL, NULL, &tv);
-  if (retval == -1) {
-    printf("error select\n");
-    return -1;
-  } else if (retval > 0) {
-    // ricevuto risposta legge 1 carattere e lo stampa
-    char c;
-    read(fd, &c, 1);
-    printf("read byte 0x%02X at address 0x%04X\n", (unsigned char)c, (unsigned int)address);
-  } else {
-    return -1;
-  }
-
-  return 0;
+  printf("read byte from address %u [x%04X]\n", (unsigned int)address, (unsigned int)address);
+  return write(fd, buff, strlen(buff));
 }
